chore: split source into poorly organized files. probably a bad idea
