+++ dorothy-manifest.rc
@@ -1,3 +1,3 @@
 #define RT_MANIFEST 24
 1 RT_MANIFEST "dorothy.exe.manifest"
-DOROTHY ICON "./src/images/dorothy.ico"
+DOROTHY ICON "./src/gui/images/dorothy.ico"
