+++ src/gui/images/coronation_ring.ico
