+++ src/gui/images/dorothy.ico
