+++ src/gui/images/intricacy_ring.ico
