+++ src/gui/images/lineage_ring.ico
